requirements.txt
# Core dependencies
pandas>=1.5.0
requests>=2.28.0
python-binance>=1.0.16
python-dotenv>=0.19.0

# Database
#sqlite3

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Export functionality
openpyxl>=3.0.10
xlsxwriter>=3.0.3
jinja2>=3.1.0

# Data processing
numpy>=1.21.0
pytz>=2022.1

# Logging and utilities
colorlog>=6.6.0
tqdm>=4.64.0

# Development dependencies (optional)
#pytest>=7.0.0
#black>=22.0.0
#flake8>=4.0.0
#mypy>=0.950

# Optional: For advanced features
#schedule>=1.1.0  # For automated runs
#cryptography>=3.4.8  # For enhanced security
